Link with -undefined dynamic_lookup
dock-support.o contains some symbols which may not be present at link
time on older OS versions.  These symbols are annotated with weak_import
and checked for at runtime, so it was possible to build on Snow Leopard
and run on Leopard, but it was not possible to build on Leopard since
these symbols are not available at link time.  This change makes it
possible to build on Leopard and places the burden of resolving those
symbols on dyld.

Resolves: https://trac.macports.org/ticket/32609

Signed-off-by: Jeremy Huddleston <jeremyhu@apple.com>
(cherry picked from commit debceb4a7fc39af77a98365a4012e988c3a3aad7)