Tell git's diff engine to recognize Objective-C source files.
The diff engine already has built-in smarts about how to annotate diffs
of Objective-C source files, but it needs to be told which files to
apply those smarts to.  (Otherwise, its default is a pattern appropriate
for C syntax.) For example, it will annotate diff hunks with headers
showing the signature of the method or the @implementation or @interface
line preceding the hunk. E.g.:

   @@ -1167,17 +1188,39 @@ - (void) someMethodName

See the gitattributes man page for details.

Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>
(cherry picked from commit d03bac215b414d69b639dc564c14bba1f37729a8)