Use "intended" frame rect to determine if window is maximized.
The intended frame is that which is queued, pending, or current, in that
order.

This fixes a problem where _NET_WM_STATE_MAXIMIZED_{HORZ,VERT} were
incorrect. Changes to the frame can be delayed.  A sequence like the
following occurred: initiate frame rect change for zoom or maximize, set
_NET_WM_STATE based in part on current frame, frame rect change actually
takes effect.  It leaves _NET_WM_STATE_MAXIMIZED_{HORZ,VERT} in the
exact opposite state as it should be.  Later, a property notify causes a
call to -update_frame, which calls -update_net_wm_state_hints, which
uses the incorrect _NET_WM_STATE to change the window's actual state.
So, the window might spontaneously maximize when it shouldn't.

Fixes <rdar://problem/10148457>

Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>