Don't drag/resize a window if we didn't see the mouse click event.
Quartz-wm had been interpreting MotionNotify events that indicated a button
was down as a drag or resize regardless of whether it had seen the ButtonPress
event.  However, the internal tracking state, like where the drag started, is
only initialized during the ButtonPress handler.  So, the window drag/resize
would be wildly off.

Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>